move relatedPosts from blocks to components
