+++ src/app/_components/RelatedPosts/index.tsx
@@ -25,7 +25,7 @@ export const RelatedPosts: React.FC<RelatedPostsProps> = props => {
         {docs?.map((doc, index) => {
           if (typeof doc === 'string') return null
 
-          return <Card key={index} doc={doc} relationTo="posts" showCategories />
+          return <Card key={index} doc={doc} relationTo="posts" showCategory />
         })}
       </div>
     </Padding>
